move demo dataviz into Demo component
